requirements.txt
sentence-transformers==2.6.1
transformers==4.40.1
torch>=1.13.0

langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.5
langchain-pinecone==0.0.3
langchain-experimental

pinecone-client[grpc]
flask
pypdf
python-dotenv

-e .
